lots of stuff

+++ Blocks/Blocks/GameObjects/Ground.cs
@@ -72,7 +72,7 @@ namespace Blocks
 
         public override void Draw(GameTime gameTime, SpriteBatch spritebach, Vector2 camera)
         {
-            spritebach.Draw(image, new Rectangle((int)Pos.X-(int)camera.X, (int)Pos.Y-(int)camera.Y, (int)BlockWidth, (int)BlockWidth), new Rectangle(0, 0, 108, 108), Color.White);
+            spritebach.Draw(image, new Rectangle((int)(Pos.X-camera.X), (int)(Pos.Y-camera.Y), (int)BlockWidth, (int)BlockWidth), new Rectangle(0, 0, 108, 108), Color.White);
         }
     }
 }
